IMPALA-2138: part 3: refactor analytic planning
This is a minor refactor of the analytic planner that moves
the logic to generate a "buffered tuple descriptor" and
equality expressions to the AnalyticEvalNode.

This is required for projection, because the buffered tuple
layout is required to match the input tuple layout, so needs
to be updated if the input tuple is projected.

After the refactoring, this can be correctly updated simply
by regenerating the tuple descriptor and expressions,
instead of attempting to "fix up" the existing expressions
with smaps, etc.

I think the plan node is actually a more approprate place for
this logic regardless, since it's tightly entwined with the
backend's assumption that the two tuple layouts are identical.

Change-Id: I691f1509a85cad51d2654278c0a4ba70a10b2706
Reviewed-on: http://gerrit.cloudera.org:8080/14325
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>