IMPALA-8996: fix test_show_create_table in test_zorder.py
IMPALA-8593 extended the filtered tbl properties in
test_show_create_table.py. This change was not propagated to
test_zorder.py which created a duplicate of the former in order
to run its tests in a custom cluster environment.

Removed the "show create table" tests from test_zorder.py.
These tests are not critical, but can cause more issues like this.
Also, Z-order tests are in a separate file only until custom
cluster is required. After merging IMPALA-8755, and removing the
feature flag (which is the reason for the custom cluster), the
tests can be merged into show-create-table.test (from
show-create-table-zorder.test).

Change-Id: Ic316224325eec64d9b1e570854a74c0372084a4a
Reviewed-on: http://gerrit.cloudera.org:8080/14359
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>