IMPALA-9005: distcc server bootstrapping improvements
Adds a script distcc_server_bootstrap.sh that can be run
in a standalone way on an ubuntu 16.04 server.

A next step could be to build a docker image that can be
deployed as an impala distcc server.

Testing:
Ran in an ubuntu 16.04 docker image and then confirmed I could
add it to my BUILD_FARM and compile things with it.

Change-Id: I9e382e90f68a5a2fe2917184fa4603bed492a308
Reviewed-on: http://gerrit.cloudera.org:8080/14363
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>