IMPALA-9008: Serialize Maven invocations to deflake impala-minimal-hive-exec
Building impala-minimal-hive-exec was flaky (sometimes it failed
to download Guava 14.0.1), and my guess is that it is related to
downloading the same dependency in parallel with ext-data-source.

Similarly to the solution in IMPALA-7051, I try to fix it by
adding a fictional dependency (ext-data-source -> shaded-deps)
to serialize the build.

Full Maven dependency graph:
fe -> yarn-extras -> ext-data-source -> shaded-deps -> impala-parent
The ordering of yarn-extras, ext-data-source, and shaded-deps are
arbitrary.

Testing:
- I never reproduced the original flakiness, so I just hope
  that it disappears with this change.

Change-Id: Ic951ad7313e5ad9add2a00e6a442c6e97f235967
Reviewed-on: http://gerrit.cloudera.org:8080/14372
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>