IMPALA-9017: Alter table events on dropped table/db should be ignored.
Alter table events received on a table/db that does not exist in
catalog can be safely ignored. This is because a table not present
in the catalog means either:
  1. The table was created in HMS but not processed by the events
     processor beacuase it is already in an error state or,
  2. It was dropped/renamed by the same catalog.

Prior to this change, alter table events generated by renames would
try to add the new table names to the cache irrespective of weather
old table was in catalog cache. After this change a drop + create
sequence happens only if table before rename exists in catalog.

Testing:
Added tests for rename + drop table/db sequences to verify events
processor does not error out.

Change-Id: I534accd704f76772ebe47865eff0fcdc7a033a48
Reviewed-on: http://gerrit.cloudera.org:8080/14385
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>