IMPALA-9034: fix distcc+ccache
According to the ccache manpage, the correct way to compose it with
distcc is to use CCACHE_PREFIX. I think this explains why local
ccache wasn't working for me.

I updated the distcc wrapper scripts to do this instead and
confirmed that it works - after doing a clean and rebuilding
the same branch, it is much faster and "ccache -s" shows a
bunch of cache hit.

ccache on the distcc server side still works.

Change-Id: Ie0b080709bd765056b9296d3deb805038fc01e5d
Reviewed-on: http://gerrit.cloudera.org:8080/14408
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>