IMPALA-8950: Set overwrite to False by default in fs_client.copy
test_udfs.py and test_load.py are consistently failing on S3 with
the following error:

E   TypeError: copy() takes exactly 4 arguments (3 given)

The issue is that a new parameter 'overwrite' was added to the method
'HadoopFsCommandLineClient.copy', but with no default value. This patch
fixes the issue by adding 'overwrite=False'.

Testing:
* Ran core tests on S3

Change-Id: I1935e0c92f7402e727ee16596485147cf1a5a516
Reviewed-on: http://gerrit.cloudera.org:8080/14432
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>