IMPALA-7504/KUDU-2979 ParseKerberosPrincipal() should use krb5_parse_name() instead
We want to use krb5_parse_name() to parse the principal instead of
using custom code.

When kerberos is initialized in Impala's copy of Kudu code, it stores a
global context which is used when accessing the Krb5 library. To use
this global context the code that parses the principal name is moved
into the Impala Kudu code. This new code is then called from the
existing ParseKerberosPrincipal method.

Test done:
Add two tests to authentication-test, one to verify parsing a principal
containing a special character. The other to verify exception when
parsing bad format principal, new error code is 2 instead of original 112
which is BAD_PRINCIPAL_FORMAT error code.
Run impala-private-parameterized tests.

Change-Id: Ifddafa7aae25d66ed7d9fa0306f17501a191cdac
Reviewed-on: http://gerrit.cloudera.org:8080/14520
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>

Change-Id: I0e64ebdc10f102dbdc5b87f6fe3f2a0310b1be24
Reviewed-on: http://gerrit.cloudera.org:8080/14433
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>