IMPALA-9052: Events processor should skip blacklisted database and table
IMPALA-8797 introduced a configuration which can blacklist certain
database and tables so that catalog never loads them. This is useful to
skip the Hive's internal book-keeping tables which are unreadable by
Impala (like Information schema and sys tables). However, if there are
events generated on such tables, the events processor goes into error
state because the database does not exist.

This patch adds support for ignoring blacklisted database and tables in
the events processor.

Testing: 1. Added a new test case which makes sure that events processor
is not in error state after receiving events on blacklisted tables and
dbs.
2. Ran MetastoreEventsProcessorTest

Change-Id: Ic5a53b722e6225e9cad9954f447f821c2c677e60
Reviewed-on: http://gerrit.cloudera.org:8080/14457
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>