Fix webserver's use of sq_printf/sq_write
Previously, the webserver used multiple calls to sq_printf and
sq_write when sending most responses.

This can lead to a bad interaction between Nagle's algorithm and TCP
delayed acks which can add significant latency to RTT.

This patch modifies the SendResponse() function to buffer the entire
response and send it in a single sq_write call.

Testing:
- Ran all existing webserver tests.

Change-Id: I343ae83d6324bc710e4cf96d66975a7c9694706f
Reviewed-on: http://gerrit.cloudera.org:8080/14471
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>