IMPALA-9055: Impala shouldn't set expiration to NEVER for cache directives.
In HdfsCachingUtil we set the expiration of cache directives to never.
This works well until the cache pool has max TTL set. Once max TTL is
set Impala will get an exception when it tries to add caching for tables
or partitions.

I changed HdfsCachingUtil to not set the expiration. This way the cache
directive inherits the expiration from the cache pool.

Testing
Added e2e test that creates a table in a cache pool that has max TTL.

Change-Id: I475b92704b19e337b2e62f766e5b978585bf6583
Reviewed-on: http://gerrit.cloudera.org:8080/14485
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>