IMPALA-9045: Filter base directories of open/aborted compactions
Base directories are in the format of base_<write_id>_<transaction_id>.
The <transaction_id> part helps to decide whether a base directory
is fully written, or is it still being written by a compaction job.
Compaction jobs don't increase the write id of a table, hence the
<write_id> part cannot be used for that.

Before this commit Impala didn't check the validity of <transaction_id>,
therefore it might read the contents of a half-written base directory.
With this change Impala retrieves the valid transaction list from HMS
and checks if <transaction_id> is committed.

Testing
 * Added an e2e test that simulates in-progress compactions
 * Added frontend test that filters based on custom valid txn list

Change-Id: Idb895df38bc075e4767e44a6887dbe3000a19ea6
Reviewed-on: http://gerrit.cloudera.org:8080/14547
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>