IMPALA-9095: Alter table events generated by renames should apply renames properly.
A recent change refactored alter table event handler for rename operations
which introduced a bug that causes renames to a different db to be applied
incorrectly to the same db. This change adds a check to see if both the old
db and the new db are in catalog and applies the change to newdb properly.

Testing:
1. We did not have a rename testcase that crossed a single db. Added a test
   case for renaming to a different db.
2. Ran MetastoreEventsProcessorTests.

Change-Id: I5bf46a75cd872aba268e15dd7c02c2cb46d64890
Reviewed-on: http://gerrit.cloudera.org:8080/14549
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>