IMPALA-9102: Have webserver fall back to --hostname
When the webserver determines its hostname, it first uses the value of
--webserver_interface. If this is a wildcard address, then it uses the
value returned by gethostname().

In some setups, gethostname() may return the short name and not the
fqdn. This is likely undesirable, eg. because this value is used to
qualify links when Knox proxying is being used, in which case the fqdn
is required.

In other places in Impala with the same basic issue, we allow users to
set the desired hostname with the --hostname flag. This patch has the
webserver fall back to the --hostname flag if --webserver_interface is
not set. If neither is set, we'll fall back to gethostname().

Testing:
- Ran existing webserver tests.
- Manually tested in an environment where gethostname() does not
  return the fqdn

Change-Id: I76d9b8b5286aa7cfc72b0ccd43de8dafda7ff9fe
Reviewed-on: http://gerrit.cloudera.org:8080/14568
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>