Add query id debugging for HS2
As part of debugging IMPALA-9098, I realised we don't print the query ID
for HS2 queries, which makes it harder finding queries in the logs.

This pulls the query ID out of the Impyla handle and prints it in
the expected form (I used the same logic as in impala shell).

Change-Id: I5d34128a7a2b69ca9043fcf5e919e9dd2a9f452b
Reviewed-on: http://gerrit.cloudera.org:8080/14607
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>