IMPALA-9090: Add name of table being scanned in scan node profile
Before this change, the only way to figure out the table being scanned
by a scan node in the profile is to pull the string out of the explain
plan or execsummary. This is awkward, both for manual and automated
analysis of the profiles. We should include the table name as a string
in the SCAN_NODE implementation.

After this change, a new line "Table Name: database.table" will be added
in first line of scan node profile.

Also fix a bug that frontend pass hbase and kudu table name incorrectly
to thrift. Before this change, native name of hbase and kudu table are
passed in and there is no way to get hms table name from TableDescriptor
in backend.

After this change, for HBaseTableDescriptor and KuduTableDescriptor,
function name() would return hms table name, function table_name()
would return hbase or kudu native table name.

Manually tested on mini-cluster with:
1. hdfs and s3 table with file format text and parquet.
2. hbase table.
3. kudu table.

Change-Id: If5da1112bcf38ae55b89eccfd7c7fad860819a99
Reviewed-on: http://gerrit.cloudera.org:8080/14660
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>